requirements.txt
Django>=4.2
mysqlclient==2.2.4
pymysql
langchain_groq
python-dotenv
python-docx
PyPDF2
gunicorn
whitenoise
dj-database-url
psycopg2-binary
